reduce boilerplate code
